feat: rotas autenticadas

+++ src/services/api-url.ts
@@ -1,10 +1,16 @@
 const baseUrl = process.env.NEXT_PUBLIC_URL_API;
+import { getCookie } from "cookies-next";
 
 export async function getUsers() {
+  const token = getCookie('dashskins-access-token') as string;
+  if (!token) {
+    return null;
+  }
   const res = await fetch(`${baseUrl}/users`, {
     method: 'GET',
     headers: {
       'Content-Type': 'application/json',
+      Authorization: `Bearer ${token}`,
     },
     cache: 'force-cache',
     next: { tags: ['users'] },
@@ -18,10 +24,16 @@ export async function getUsers() {
 };
 
 export async function createUser(data: { username: string, email: string, age: number, avatar: string }) {
+  const token = getCookie('dashskins-access-token') as string;
+  if (!token) {
+    return null;
+  }
+
   const res = await fetch(`${baseUrl}/users`, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
+      Authorization: `Bearer ${token}`,
     },
     body: JSON.stringify(data),
   });
@@ -34,10 +46,16 @@ export async function createUser(data: { username: string, email: string, age: n
 };
 
 export async function editUser(data: { _id: string, username: string, email: string, age: number, avatar: string }) {
+  const token = getCookie('dashskins-access-token') as string;
+  if (!token) {
+    return null;
+  }
+
   const res = await fetch(`${baseUrl}/users/${data._id}`, {
     method: 'PUT',
     headers: {
       'Content-Type': 'application/json',
+      Authorization: `Bearer ${token}`,
     },
     body: JSON.stringify(data),
   });
@@ -50,10 +68,16 @@ export async function editUser(data: { _id: string, username: string, email: str
 };
 
 export async function deleteUser(id: string) {
+  const token = getCookie('dashskins-access-token') as string;
+  if (!token) {
+    return null;
+  }
+
   const res = await fetch(`${baseUrl}/users/${id}`, {
     method: 'DELETE',
     headers: {
       'Content-Type': 'application/json',
+      Authorization: `Bearer ${token}`,
     },
   });
